Debug library path in init.py
For developper mode, before this commit, it failed because the path is not in site-package.
Now, if developper mode (ie. pip install -e .), it does not look if the path are correct

For LD_PRELOAD, check if the 2 librairies are loaded but now it does not care about the which library is loaded first in the path