Phs source fixed segmentation fault (#328)
Fixed segmentation faults due to wrong variable type when passing from python to cpp. Was double, should be float

* Enforce correct data type on python side (in case different data types in root file or delivered via uproot

* restructured data transfer from python to cpp

* minor adjustments

* added method to rotate the phase space source in addition to movement required by treatment plan for treatment plan phs source

* fixed wrongly initialized value for entry_start for multithreaded mode

* fixed initialization for multithreading of phs source

* Update phspsources.py

In multithreading mode
entry_start needs to be an array containing one entry for every thread.
If it is not created by the user,
create a entry_start array with the correct number of start entries
all entries are spaced by the ceil(number of particles/thread)+1

* added improved initialization of starting values for phase space files and sources in multi threading

* Improved phase space source documentation

---------

Co-authored-by: Thomas BAUDIER <thomas.baudier@creatis.insa-lyon.fr>