Raise an error with material containing an element with a fraction of 0
In windows only, a material with an element with a fraction at 0 lead to a infinite loop in G4 during the simlulation
To avoid the error, I raise an error to force the user to have material without element with a fraction of 0.

Update tests using a GateMaterials.db with elements containing f=0. If it was possible I prefer to use the generic .db instead of the simu one