Proof of concept on generic filters (see long part)
The entrypoint is in the base actor interface, using filter member.
This string can be defined to a Python expression like:

"particle_name == "gamma" and pre_kinetic_energy <= 20 * MeV"

A Transformer will:
- pre compute constants (e.g. 20 * MeV will become a direct value);
- turn attribute names into function calls to acquire its value.

The AST is then used as the body of a Python function.
This function is given to the C++-side corresponding base actor to be
used as the filter.

This POC works, though it is noticeably slower than another experiment
using a JIT compiled C++ function (using ROOT.gInterpreter).