Update materials and translations in macaco.py
Thanks for the review!
I updated the default camera position to be at the world center, as suggested and fixed the coordinate of pcb_scatt. For the material, I checked the G4 files and it's not the same at all, I'm not sure why I used it in my code before, so I fixed it here and on vs.